consoliadte Enum case test, close #570
